requirements.txt
python-telegram-bot~=21.11.1
requests~=2.32.3
Flask~=3.1.0
python-dotenv~=1.0.1
PyMySQL~=1.1.1
pandas~=2.2.3